LayoutTests/inspector/debugger/network-uisourcecode-provider.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/workspace-test.js"></script>
<script>
function test()
{
    var mockContentsMap = {};
    var lastResouceId = 1;

    NetworkAgent.getResponseBody = function(requestId, callback)
    {
        callback(undefined, mockContentsMap[requestId], false);
    }

    PageAgent.getResourceContent = function(frameId, url, callback)
    {
        callback(undefined, mockContentsMap[url], false);
    }

    function createResourceMock(type, withPendingRequest, content)
    {
        var documentURL = "http://fake.url";
        var resourceId = ++lastResouceId + "";
        var url = documentURL + "/" + resourceId;
        var frameId = "frame-id";
        var loaderId = "loader-id";
        var mimeType;
        switch (type) {
        case WebInspector.resourceTypes.Document:
            mimeType = "text/html";
            break;
        case WebInspector.resourceTypes.Script:
            mimeType = "text/javascript";
            break;
        case WebInspector.resourceTypes.Stylesheet:
            mimeType = "text/css";
            break;
        }

        if (withPendingRequest) {
            var request = new WebInspector.NetworkRequest(resourceId, url, "http://fake.url", frameId, loaderId);
            request.type = type;
            request.mimeType = mimeType;
            mockContentsMap[resourceId] = content;
        }
        var resource = new WebInspector.Resource(request, url, documentURL, frameId, loaderId, type, mimeType);
        mockContentsMap[url] = content;
        WebInspector.resourceTreeModel.dispatchEventToListeners(WebInspector.ResourceTreeModel.EventTypes.ResourceAdded, resource);

        return resource;
    }

    function finishResource(resource)
    {
        resource.request.finished = true;
        resource.request.dispatchEventToListeners(WebInspector.NetworkRequest.Events.FinishedLoading, resource.request);
    }

    function createNetworkUISourceCodeProvider()
    {
        InspectorTest.createWorkspace();
        WebInspector.resourceTreeModel = new WebInspector.Object();
        var networkUISourceCodeProvider = new WebInspector.NetworkUISourceCodeProvider(InspectorTest.testWorkspace);
        return networkUISourceCodeProvider;
    }

    InspectorTest.runTestSuite([
        function testDocumentResourceWithPendingRequest(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();

            InspectorTest.addResult("Creating request.");
            var resource = createResourceMock(WebInspector.resourceTypes.Document, true, "<document request content>");

            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            InspectorTest.addResult("Finishing request.");
            finishResource(resource);

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testDocumentResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            var resource = createResourceMock(WebInspector.resourceTypes.Document, false, "<document resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testScriptResourceWithPendingRequest(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();

            InspectorTest.addResult("Creating request.");
            var resource = createResourceMock(WebInspector.resourceTypes.Script, true, "<script request content>");

            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            InspectorTest.addResult("Finishing request.");
            finishResource(resource);

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testScriptResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            var resource = createResourceMock(WebInspector.resourceTypes.Script, false, "<script resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testStylesheetResourceWithPendingRequest(next)
        {
            var endTestCallsLeft = 2;
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            InspectorTest.addResult("Creating request.");
            var resource = createResourceMock(WebInspector.resourceTypes.Stylesheet, true, "<stylesheet request content>");
            InspectorTest.addResult("Finishing request.");
            finishResource(resource);
            endTest();

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, endTest);
            }

            function endTest()
            {
                if (--endTestCallsLeft)
                    return;
                next();
            }
        },

        function testStylesheetResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            var resource = createResourceMock(WebInspector.resourceTypes.Stylesheet, false, "<stylesheet resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests NetworkUISourceCodeProvider class.</p>
</body>
</html>
